refactor(struct): isolate generic consultation receipt boundary
