+++ src/struct/index.ts
@@ -7,10 +7,6 @@ export {
   type StructDocumentJson,
 } from './codec'
 export * from './ids'
-export {
-  validateModelConsultationReceipt,
-  type ModelFallbackReceipt,
-} from './model-consultation-receipt'
 export * from './reading-order'
 export * from './recovery'
 export * from './xhtml'
